feat(web/title): move to separate component
